Mv some funcs to tools
